Decompose AnalysisView
